agent-examples/01-multi-agent-mixed-repo/requirements.txt
python-dotenv
langchain-core
langchain-openai
langgraph
openai
